requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==24.2
graphene-django==3.2.0
django-graphql-jwt==0.3.2
python-dotenv==1.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
pytest==8.0.0
pytest-django==4.8.0
camelot-py==1.0.9
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pdfplumber==0.11.7
PyMuPDF==1.26.5
pytesseract==0.3.13
